move use-duplicate-dashcard to better directory
